backend/requirements.txt
# NiksES Backend Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Email Parsing
mail-parser==3.15.0
extract-msg==0.48.4

# HTTP Client
httpx==0.27.0
aiohttp==3.9.3

# DNS & WHOIS
dnspython==2.6.1
python-whois==0.9.4

# QR Code Detection
pyzbar==0.1.9
Pillow==10.2.0

# Security
cryptography==42.0.2

# AI
openai==1.12.0

# Utilities
python-dateutil==2.8.2
validators==0.22.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
python-dotenv>=1.0.0
reportlab>=4.0.0
